% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_diversity_from_transition}
\alias{get_diversity_from_transition}
\title{Compute the diversity of a transition matrix, that is the individual diversities of
its rows or their mean diversity.}
\usage{
get_diversity_from_transition(transition, type = "individual",
  mean_distribution = NULL, order = NULL, measure = NULL)
}
\arguments{
\item{transition}{A matrix of floats in [\code{0},\code{1}], with all lines summing to
\code{1}, giving the transition matrix from which the individual diversity values are
computed.}

\item{type}{Either \code{'individual'}, to separately compute all individual diversities,
or \code{'mean'}, to compute their geometric mean.}

\item{mean_distribution}{(optional, only when \code{type == 'mean'}) A vector of floats
in [\code{0},\code{1}] and summing to \code{1} giving the probability distribution that
is used to weight the diversity values when computing their geometric means. It should
hence contain as many values as there are rows in the input \code{transition}. If not
specified, this distribution is assumed uniform.}

\item{order}{A vector of positive floats (possibly including \code{Inf}) giving the orders
of the diversity measures to be computed. If neither \code{order} nor \code{measure} is
specified, a predefined list of 8 diversity measures is computed.}

\item{measure}{A vector of strings giving the names of the diversity measures to compute.
Possible values are \code{richness}, \code{entropy}, \code{herfindahl}, and
\code{bergerparker}.}
}
\value{
A matrix (or a vector) of positive floats giving the individual diversity values
(or their geometric mean) of the row of the input \code{transition} matrix.
}
\description{
\code{get_diversity_from_transition} either computes the individual diversity values
associated to the rows of the input \code{transition} matrix, or the geometric means of
these values, optionally weighting them according to the input \code{mean_distribution}.
This hence allows to compute conditional diversity values associated to a transition
matrix.
}
\examples{
transition <- matrix (c (1/3, 1/3, 1/3, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)

get_diversity_from_transition (transition, type='individual', order=c(0,Inf), measure='entropy')
get_diversity_from_transition (transition, type='mean')
get_diversity_from_transition (transition, type='mean', mean_distribution=c(1/4,3/4))

}
