% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_multipartite}
\alias{get_multipartite}
\title{Build a properly-structured multipartite graph from raw data.}
\usage{
get_multipartite(filename = NULL, data = NULL)
}
\arguments{
\item{filename}{A character string giving the path to the file containing the raw data.

The input file should contain at least four columns, separated by spaces. Each row gives
a link between two nodes belonging to two different parts of the multipartite graph.
The first column gives the label of the part of the first node and the second column
gives the label of this node (any character string). Similarly, the third and fourth
columns give the labels of the second part and of the second node. The file can eventually
contain a fifth column giving the weights of the links (any positive integer or float
value).}

\item{data}{A \code{data.frame} containing the raw data.

This \code{data.frame} should have the same structure than the one described above
for the case of an input file: four columns indicating the labels of the
parts and of the nodes constituting the link, and an optional
fifth column for its weight. At least \code{filename} or \code{data} should be
specified, but not both at the same time.}
}
\value{
A properly-structured multipartite graph that can be used by the other functions
of the \code{triversity} package.

The resulting object encodes the labels of the parts and of the nodes in the multipartite
graph, as well as the transition matrices of random walks following different paths
between parts (encoded as sparse matrices of floats in [\code{0},\code{1}], with all
rows summing to \code{1}). These transition matrices
are then used by functions such as \code{\link{get_distribution_from_path}} to
compute the probability distributions of such random walks, or such as
\code{\link{get_diversity_from_path}} to compute the diversity of these distributions.

Assuming the object returned by \code{get_multipartite} is stored in the
\code{graph} variable, then:
\itemize{
\item \code{graph$parts} is a vector of character strings giving the labels of the
different parts in the multipartite graph;
\item \code{graph$nodes} is a list of vectors of character strings giving the labels
of the nodes constituting the different parts of the multipartite graph;
\item \code{graph$edges} is a \code{data.tree} whose nodes each contains the
transition matrix of the corresponding random walk.
}
}
\description{
\code{get_multipartite} builds a properly-structured multipartite graph
from a file or from a \code{data.frame} containing raw data. This object can then
be used by the other functions of this package. The structure
of the input data and of the resulting structure is detailed below.
}
\examples{
data (tripartite_example)
graph <- get_multipartite (data=tripartite_example)

graph$parts

part1 <- graph$parts[1]
graph$nodes[[part1]]

part2 <- graph$parts[2]
as.matrix (graph$edges$Climb(part1,part2)$matrix)

}
\seealso{
Package \code{slam} for the sparse encoding of transition matrices and
package \code{data.tree} for the tree structure storing these matrices.
}
