% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_transition_from_path}
\alias{get_transition_from_path}
\title{Get a probability transition matrix from a path in the tripartite graph}
\usage{
get_transition_from_path(tripartite, path)
}
\arguments{
\item{path}{The path to follow to compute the transition matrix}
}
\value{
The resulting transition matrix
}
\description{
Get a probability transition matrix from a path in the tripartite graph
}
