% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\docType{data}
\name{tripartite_example}
\alias{tripartite_example}
\title{An example of dataframe encoding a small tripartite graph.}
\format{An object of class \code{data.frame} with 11 rows and 4 columns.}
\usage{
data(tripartite_example)
}
\description{
\code{tripartite_example} is a \code{data.frame} containing raw data that encodes a small
tripartite graph. It has the proper format to be loaded by \code{\link{get_multipartite}}.
It contains four columns. Each row gives a link between two nodes belonging to two
different parts of the tripartite graph. The first column gives the label if the part of
the first node and the second column gives the label of this node (any character string).
Similarly, the third and fourth columns give the labels of the second part and of the
second node. A fifth column could eventually be added to give the weights of the links
(any positive integer or float value).
}
\examples{
data (tripartite_example)
head (tripartite_example)

graph <- get_multipartite (data=tripartite_example)
graph
}
\keyword{datasets}
