% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\docType{package}
\name{triversity}
\alias{triversity}
\alias{triversity-package}
\title{Compute diversity measures on tripartite graphs.}
\description{
\code{triversity} is an R package for the computation of diversity measures
on tripartite graphs. It first implements a parametrised family of such
diversity measures applying on probability distributions. Sometimes called
"True Diversity", this family contains famous measures such as the Richness,
the Shannon entropy, the Herfindahl-Hirschman index, and the Berger-Parker
index. This package then allows to apply such measures on probability
distributions resulting from random walks on tripartite graphs. By defining
an initial distribution at a given level in the graph, a path within the three
levels, and eventually a conditional step, the probability of the walker's
position within the final level is then computed, thus providing a particular
instance of diversity index.
}
\details{
This package has been developed by researchers of the Complex Networks team,
(http://www.complexnetworks.fr/) within the Computer Science Laboratory of Paris 6
(https://www.lip6.fr/), for the AlgoDiv project (http://algodiv.huma-num.fr/),
founded by the French National Agency of Research
(http://www.agence-nationale-recherche.fr/) under grant ANR-15-CE38-0001.

Contact: Robin Lamarche-Perrin <Robin.Lamarche-Perrin@lip6.fr>

See also my webpage: https://www-complexnetworks.lip6.fr/~lamarche/

List of main collaborators:
\itemize{
\item Lionel Tabourier
\item Fabien Tarissan
\item Rapha\"el Fournier S'niehotta
\item R\'emy Cazabet
}

Copyright \copyright 2017 Robin Lamarche-Perrin

\code{triversity} is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version. It is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with this program. If not, see <http://www.gnu.org/licenses/>.
}
