% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_conditional_diversity_from_distribution}
\alias{get_conditional_diversity_from_distribution}
\title{Get the conditional diversity associated to a probability distribution and a transition matrix}
\usage{
get_conditional_diversity_from_distribution(distribution, transition,
  order = NULL, measure = NULL)
}
\arguments{
\item{distribution}{The probability distribution to measure}

\item{distribution}{The probability transition matrix to use}
}
\value{
The diversity of the input distribution
}
\description{
Get the conditional diversity associated to a probability distribution and a transition matrix
}
