% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_distribution_from_path}
\alias{get_distribution_from_path}
\title{Get a probability distribution from a path in the tripartite graph}
\usage{
get_distribution_from_path(tripartite, path, initial_node = NULL,
  initial_distribution = NULL)
}
\arguments{
\item{path}{The path to follow to compute the distribution}
}
\value{
The resulting probability distribution
}
\description{
Get a probability distribution from a path in the tripartite graph
}
