% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_diversity_from_distribution}
\alias{get_diversity_from_distribution}
\title{Get the diversity associated to a probability distribution}
\usage{
get_diversity_from_distribution(distribution, order = NULL, measure = NULL)
}
\arguments{
\item{distribution}{The probability distribution to measure}
}
\value{
The diversity of the input distribution
}
\description{
Get the diversity associated to a probability distribution
}
