% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_diversity_from_path}
\alias{get_diversity_from_path}
\title{Get the diversity associated to the distribution obtained through a path in the loaded tripartite graph}
\usage{
get_diversity_from_path(tripartite, path, conditional_path = NULL,
  initial_node = NULL, initial_distribution = NULL, order = NULL,
  measure = NULL)
}
\arguments{
\item{path}{The path to follow to build the probability distribution}
}
\value{
The diversity of the resulting probability distribution
}
\description{
Get the diversity associated to the distribution obtained through a path in the loaded tripartite graph
}
