% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triversity.R
\name{get_tripartite}
\alias{get_tripartite}
\title{Load a tripartite graph}
\usage{
get_tripartite(filename = NULL, data = NULL)
}
\arguments{
\item{filename}{The file containing the tripartite graph}

\item{data}{A data.frame containing the tripartite graph}
}
\value{
A tripartite graph
}
\description{
Load a tripartite graph
}
